+++ internal/providers/aws/api.go
@@ -3,6 +3,11 @@ package aws
 import (
 	"context"
 	"errors"
+	"fmt"
+	cloudwatchType "github.com/aws/aws-sdk-go-v2/service/cloudwatch/types"
+	ec2Types "github.com/aws/aws-sdk-go-v2/service/ec2/types"
+	"github.com/aws/aws-sdk-go/aws/awserr"
+	"log"
 	"strconv"
 	"time"
 
@@ -227,6 +232,198 @@ func IsValidMonth(month string) bool {
 		return false
 	}
 	return true
+}
+func (p *AWSCloud) DescribeRegions(ctx context.Context, param types.DescribeRegionsRequest) (types.DescribeRegions, error) {
+	input := &ec2.DescribeRegionsInput{}
+	response, err := p.ec2Client.DescribeRegions(ctx, input)
+	if err != nil {
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			default:
+				log.Println(aerr.Error())
+			}
+		} else {
+			log.Println(err.Error())
+		}
+		return types.DescribeRegions{}, err
+	}
+	if response.Regions != nil {
+		itemRegions := make([]types.ItemRegion, 0, len(response.Regions))
+		for _, regin := range response.Regions {
+			newRegion := types.ItemRegion{
+				RegionId:  aws.StringValue(regin.RegionName),
+				LocalName: _regionLocalName[aws.StringValue(regin.RegionName)],
+			}
+			itemRegions = append(itemRegions, newRegion)
+		}
+		return types.DescribeRegions{
+			List: itemRegions,
+		}, err
+	}
+	return types.DescribeRegions{}, err
+}
+
+func (p *AWSCloud) DescribeInstances(ctx context.Context, param types.DescribeInstancesRequest) (types.DescribeInstances, error) {
+	input := &ec2.DescribeInstancesInput{
+		InstanceIds: param.InstanceIds,
+	}
+	output, err := p.ec2Client.DescribeInstances(ctx, input)
+	if err != nil {
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			default:
+				log.Println(aerr.Error())
+			}
+		} else {
+			log.Println(err.Error())
+		}
+	}
+	if output.Reservations != nil {
+		reservedInstances, err1 := p.describeReservedInstances(ctx)
+		if err1 != nil {
+			log.Println(err1.Error())
+			return types.DescribeInstances{}, err1
+		}
+		return convDescribeInstances(output.Reservations, reservedInstances), err
+	}
+	return types.DescribeInstances{}, err
+}
+
+//Get Reserved Instances
+func (p *AWSCloud) describeReservedInstances(ctx context.Context) (map[string]string, error) {
+	reservedInstances := make(map[string]string, 0)
+	output, err := p.ec2Client.DescribeReservedInstances(ctx, &ec2.DescribeReservedInstancesInput{})
+	if err != nil {
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			default:
+				log.Println(aerr.Error())
+			}
+		} else {
+			log.Println(err.Error())
+		}
+		return reservedInstances, err
+	}
+	if output.ReservedInstances != nil {
+		for _, reservation := range output.ReservedInstances {
+			if ec2Types.ReservedInstanceStateActive == reservation.State {
+				reservedInstances[aws.StringValue(reservation.ReservedInstancesId)] = string(reservation.InstanceType)
+			}
+		}
+	}
+	return reservedInstances, err
+}
+
+func convDescribeInstances(reservations []ec2Types.Reservation, reservedInstances map[string]string) types.DescribeInstances {
+	awsInstances := make([]types.ItemDescribeInstance, 0)
+	for _, reservation := range reservations {
+		for _, instance := range reservation.Instances {
+			region := aws.StringValue(instance.Placement.AvailabilityZone)
+			subscriptionType := cloud.PostPaid
+			for k, v := range reservedInstances {
+				if string(instance.InstanceType) == v {
+					subscriptionType = cloud.PrePaid
+					delete(reservedInstances, k)
+				}
+			}
+			newInstance := types.ItemDescribeInstance{
+				InstanceId:       aws.StringValue(instance.InstanceId),
+				InstanceName:     aws.StringValue(instance.Tags[0].Value),
+				RegionId:         aws.StringValue(instance.Placement.AvailabilityZone),
+				RegionName:       _regionLocalName[region[0:len(region)-1]],
+				SubscriptionType: subscriptionType,
+				PublicIpAddress:  []string{aws.StringValue(instance.PublicIpAddress)},
+				InnerIpAddress:   []string{aws.StringValue(instance.PrivateIpAddress)},
+			}
+			awsInstances = append(awsInstances, newInstance)
+		}
+	}
+	return types.DescribeInstances{
+		TotalCount: len(awsInstances),
+		List:       awsInstances,
+	}
+}
+
+func convDescribeMetricListRequest(param types.DescribeMetricListRequest) (*cloudwatch.GetMetricDataInput, map[string]string) {
+	ids := make(map[string]string)
+	metricDataQueries := []cloudwatchType.MetricDataQuery{}
+	var nameSpace, metricName, label string
+	if types.MetricItemCPUUtilization == param.MetricName {
+		nameSpace = Namespace_Cpu
+		metricName = CPUUtilization
+		label = CPUUtilization
+	}
+	if types.MetricItemMemoryUsedUtilization == param.MetricName {
+		nameSpace = Namespace_Mem
+		metricName = MemoryUtilization
+		label = MemoryUtilization
+	}
+	period, _ := strconv.Atoi(param.Period)
+	for i, instanceId := range param.Filter.InstanceIds {
+		dimension := cloudwatchType.Dimension{
+			Name:  aws.String(InstanceId),
+			Value: aws.String(instanceId),
+		}
+		metricDataQuery := cloudwatchType.MetricDataQuery{
+			Id: aws.String(fmt.Sprintf("%s%s", "instance", strconv.Itoa(i))),
+			MetricStat: &cloudwatchType.MetricStat{
+				Metric: &cloudwatchType.Metric{
+					Namespace:  aws.String(nameSpace),
+					MetricName: aws.String(metricName),
+					Dimensions: []cloudwatchType.Dimension{dimension},
+				},
+				Stat:   aws.String(string(cloudwatchType.StatisticAverage)),
+				Period: aws.Int32(int32(period)),
+			},
+			Label: aws.String(label),
+		}
+		metricDataQueries = append(metricDataQueries, metricDataQuery)
+		ids[aws.StringValue(metricDataQuery.Id)] = instanceId
+	}
+	input := &cloudwatch.GetMetricDataInput{
+		StartTime:         aws.Time(param.StartTime),
+		EndTime:           aws.Time(param.EndTime),
+		MetricDataQueries: metricDataQueries,
+	}
+	return input, ids
+}
+func (p *AWSCloud) DescribeMetricList(ctx context.Context, param types.DescribeMetricListRequest) (types.DescribeMetricList, error) {
+	if param.Filter.InstanceIds == nil || len(param.Filter.InstanceIds) == 0 {
+		return types.DescribeMetricList{}, nil
+	}
+	request, ids := convDescribeMetricListRequest(param)
+	output, err := p.cloudWatch.GetMetricData(ctx, request)
+	if err != nil {
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			default:
+				log.Println(aerr.Error())
+			}
+		} else {
+			log.Println(err.Error())
+		}
+		return types.DescribeMetricList{}, err
+	}
+	if output.MetricDataResults != nil {
+		list := make([]types.MetricSample, 0)
+		for _, metricDataResult := range output.MetricDataResults {
+			if len(metricDataResult.Values) > 0 {
+				for i, value := range metricDataResult.Values {
+					metricSample := types.MetricSample{
+						InstanceId: ids[aws.StringValue(metricDataResult.Id)],
+						Average:    value,
+						Timestamp:  aws.TimeUnixMilli(metricDataResult.Timestamps[i]),
+					}
+					list = append(list, metricSample)
+				}
+			}
+
+		}
+		return types.DescribeMetricList{
+			List: list,
+		}, err
+	}
+	return types.DescribeMetricList{}, err
 }
 
 func (p *AWSCloud) DescribeInstanceBill(ctx context.Context, param types.DescribeInstanceBillRequest, isAll bool) (types.DescribeInstanceBill, error) {
